+++ tests/ui/pinned_buffered_unpin_fail.rs
@@ -1,11 +1,10 @@
 use std::marker::PhantomPinned;
-use std::ops::Deref;
 use std::pin::Pin;
 
 use dynify::{from_closure, Buffered, Dynify};
 
 // Wraps Pin::into_inner to prevent rustc from reporting errors from rust-src.
-fn unpin<P: Deref>(ptr: Pin<P>) -> P
+fn unpin<P: std::ops::Deref>(ptr: Pin<P>) -> P
 where
     P::Target: Unpin,
 {
@@ -13,7 +12,7 @@ where
 }
 
 fn main() {
-    let mut stack = [0u8; 16];
+    let mut stack = std::mem::MaybeUninit::<[u8; 16]>::uninit();
     let init = from_closure(|slot| slot.write(PhantomPinned));
     let val: Buffered<PhantomPinned> = init.init(&mut stack);